+++ CHANGELOG.md
@@ -7,6 +7,9 @@ Documentation for TransferBench is available at
 ### Added
 - Added USE_HSA_DMA env var to switch to using hsa_amd_memory_async_copy instead of hipMemcpyAsync for DMA execution
 - Added ability to set USE_GPU_DMA env var for a2a benchmark
+- Adding check for large BAR enablement for GPU devices during topology check
+### Fixed
+- Potential memory leak if HSA reports 0 hops between GPUs and CPUs
 
 ## v1.51